First version
Code has two main functions, both take a base double and an exponent integer. They return the base to the power of the exponent. One function is written recursively, and one iteratively. The time each run in nanoseconds is recorded for both functions for exponent values 0 to 35,000. This data is output to a .csv file.